Organize gnupg in package
